+++ spec/features/users_can_create_projects_spec.rb
@@ -25,6 +25,8 @@ RSpec.feature "Users can create new projects" do
         fill_in "Year", with: completion_date.year
       end
 
+      fill_in "School SharePoint link", with: "https://educationgovuk-my.sharepoint.com/project/information"
+
       within("#advisory-board-date") do
         fill_in "Day", with: two_weeks_ago.day
         fill_in "Month", with: two_weeks_ago.month