+++ CHANGELOG.md
@@ -12,6 +12,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   conditions.
 - Users can be assigned to projects roles via 'change' links on the project
   information page. Only users with the team lead role can assign users.
+- The Handover/new project form collects an establishment SharePoint link.
+- The Project Details section of the Project Information shows the establishment
+  SharePoint link.
+- The Project Summary shows the establishment SharePoint link.
 
 ### Changed